Add -E to use pre-processor on input before generating bitcode.
This can be useful in situations where a locally customized "native" compiler is
being used, where the customization includes modifications to header files and
internal macro definition values that will not be processed successfully by a
general clang compiler installation which has different include files and builtin
definitions (including #ifdef __gnu__ v.s. #ifdef __clang__).

Using this flag, the native compiler is used for pre-processing only, and then
the result of that pre-processing is provided to the clang compiler for bitcode
generation.  This helps ensure that all pre-processing (include files,
definitions, etc.) are handled by the native compiler and that clang is provided
with a single, coherent source file to compile.

There are some low-level differences between gcc and clang that can be
problematic (e.g. #ifdef __gnu__ protected sections); these usually define
low-level operations like atomics differently than clang defines them (and
incompatibly).  The build-bom generate-bitcode operation internally provides
workarounds to avoid some of these, and the --inject-argument can be used to
address others that are encountered when using the -E flag for pre-processing.