Add ability to use native compiler to pre-process the input files.
Some configurations use a compiler (or cross-compiler) that has been heavily
customized with additional command-line flags, internal definitions, and custom
include files.  If this original compiler is *not* clang then it is typically not
possible to run the same compilation request on clang to obtain llvm bitcode.

This change provides a flag that will perform the pre-processing stage (only)
with the native compiler and then pass the result of the pre-processing to clang
for obtaining the bitcode.  This usually avoids the above issues because the
pre-processing tends to handle all the customizations and the resulting C code is
fairly generic.

There are a couple of compiler directives regarding atomics and malloc resource
management that *are* different between gcc and clang, so when this mode is used,
build-bom internally provides definitions for those that elide those from the
code being used to generate bitcode (unless --strict mode is used, in which case
the user must handle these explicitly with --inject-argument arguments).  This
elision is usually acceptable because the llvm bitcode is not being used to
generate the final executable and the users of the bitcode are not concerned with
these fine-grained directives.